调式接口

+++ joybar/src/main/java/com/shenma/yueba/baijia/modle/MyFavoriteProductListPic.java
@@ -11,17 +11,17 @@ import java.io.Serializable;
 public class MyFavoriteProductListPic implements Serializable{
 
 	String pic="";// 图片地址
-    double Ratio;// 0.7500   图片高度/图片宽度的比例
+    float Ratio;// 0.7500   图片高度/图片宽度的比例
 	public String getPic() {
 		return pic;
 	}
 	public void setPic(String pic) {
 		this.pic = pic;
 	}
-	public double getRatio() {
+	public float getRatio() {
 		return Ratio;
 	}
-	public void setRatio(double ratio) {
+	public void setRatio(float ratio) {
 		Ratio = ratio;
 	}
